MainActivity view finished

+++ Assignment3/app/src/main/java/com/example/yelpit/BusinessResult.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 //GET: /businesses/search
 data class BusinessResult(
-    @SerializedName("rating") val rating: Double,
+    @SerializedName("rating") val rating: Float,
     @SerializedName("id") val id: String,
     @SerializedName("is_closed") val isClosed: Boolean,
     @SerializedName("review_count") val reviewCount: Int,
